Ray-sphere intersection

+++ RayTracing/src/Ray.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Vec3.h"
+#include "Vec3.h"..
 
 class ray {
 public:
@@ -11,7 +11,7 @@ public:
 	point3 origin() const { return orig; }
 	vec3 direction() const { return dir; }
 
-	point3 at(float t) const {
+	point3 at(double t) const {
 		return orig + t * dir;
 	}
