+++ RayTracing/src/Vec3.h
@@ -9,15 +9,15 @@ class vec3 {
 public:
 	// Constructors
 	vec3() : e{ 0,0,0 } {}
-	vec3(float e0, float e1, float e2) : e{ e0, e1, e2 } {}
+	vec3(double e0, double e1, double e2) : e{ e0, e1, e2 } {}
 
-	float x() const { return e[0]; }
-	float y() const { return e[1]; }
-	float z() const { return e[2]; }
+	double x() const { return e[0]; }
+	double y() const { return e[1]; }
+	double z() const { return e[2]; }
 
 	vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }
-	float operator[](int i) const { return e[i]; }
-	float& operator[](int i) { return e[i]; }
+	double operator[](int i) const { return e[i]; }
+	double& operator[](int i) { return e[i]; }
 
 	vec3& operator+=(const vec3& v) {
 		e[0] += v[0];
@@ -26,27 +26,27 @@ public:
 		return *this;
 	}
 
-	vec3& operator*=(const float t) {
+	vec3& operator*=(const double t) {
 		e[0] *= t;
 		e[1] *= t;
 		e[2] *= t;
 		return *this;
 	}
 
-	vec3& operator/=(const float t) {
+	vec3& operator/=(const double t) {
 		return *this *= 1 / t;
 	}
 
-	float length() const {
+	double length() const {
 		return sqrt(length_squared());
 	}
 
-	float length_squared() const {
+	double length_squared() const {
 		return e[0] * e[0] + e[1] * e[1] + e[2] * e[2];
 	}
 
 private:	// This might need to change
-	float e[3];
+	double e[3];
 };
 
 // Type aliases for vec3
@@ -71,19 +71,19 @@ inline vec3 operator*(const vec3& u, const vec3& v) {
 	return vec3(u[0] * v[0], u[1] * v[1], u[2] * v[2]);
 }
 
-inline vec3 operator*(float t, const vec3& v) {
+inline vec3 operator*(double t, const vec3& v) {
 	return vec3(t * v[0], t * v[1], t * v[2]);
 }
 
-inline vec3 operator*(const vec3& v, float t) {
+inline vec3 operator*(const vec3& v, double t) {
 	return t * v;
 }
 
-inline vec3 operator/(vec3 v, float t) {
+inline vec3 operator/(vec3 v, double t) {
 	return (1 / t) * v;
 }
 
-inline float dot(const vec3& u, const vec3& v) {
+inline double dot(const vec3& u, const vec3& v) {
 	return u[0] * v[0]
 		+ u[1] * v[1]
 		+ u[2] * v[2];
